requirements.txt

discord.py
aiosqlite
flask
